+++ Modules/FSRMDsc/DSCResources/DSR_FSRMFileScreenTemplateAction/DSR_FSRMFileScreenTemplateAction.psm1
@@ -677,7 +677,7 @@ function Test-TargetResource
             } # if
 
             if (($PSBoundParameters.ContainsKey('ReportTypes')) `
-                    -and ($action.ReportTypes -ne $ReportTypes))
+                    -and (-not (Compare-Object -ReferenceObject $action.ReportTypes -DifferenceObject $ReportTypes)))
             {
                 Write-Verbose -Message ( @(
                         "$($MyInvocation.MyCommand): "